+++ src/Components/ExecuteStrategy/interface.ts
@@ -1,0 +1,5 @@
+export interface ITradeIntent {
+  value: string;
+  assetFrom: string;
+  assetTo: string;
+}
